requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
PyMySQL==1.1.0
cryptography==41.0.4
email-validator==2.0.0
python-dotenv==1.0.0
itsdangerous==2.1.2
Werkzeug==2.3.7
Pillow==10.1.0
openpyxl==3.1.5
